requirements.txt
pygame==2.1.2
numpy==1.23.4

--extra-index-url https://download.pytorch.org/whl/cu117
torch
torchvision
torchaudio
Pillow==9.3.0
matplotlib==3.6.2